move DeferredPromise to API
